+++ app/home/open-request/template.hbs
@@ -1,56 +1,2 @@
 {{page-title 'OpenRequest'}}
-<Header>Open new request</Header>
-<div class='container'>
-  <div class='row'>
-    <div class='col'>
-      <div class='form-floating mb-3' {{on 'input' this.onTypeChange}}>
-        <select
-          class='form-select'
-          aria-label='Default select example'
-          value={{@type}}
-        >
-          <option selected>Select request type</option>
-          <option value='1.0 tpm'>1.0 TPM</option>
-          <option value='2.0 fault'>2.0 Fault</option>
-          <option value='3.0 breakdown'>3.0 Breakdown</option>
-        </select>
-      </div>
-    </div>
-    <div class='col'>
-      <div class='form-floating mb-3' {{on 'input' this.onMachineChange}}>
-        <select
-          class='form-select'
-          aria-label='Default select example'
-          value={{@machine}}
-        >
-          <option selected>Select machine</option>
-          <option value='Machine 1'>Machine 1</option>
-          <option value='Machine 2'>Machine 2</option>
-          <option value='Machine 3'>Machine 3</option>
-        </select>
-      </div>
-    </div>
-  </div>
-  <div class='form-floating'>
-    <input
-      type='text'
-      class='form-control'
-      id='floatingtext'
-      placeholder='text'
-      value={{@title}}
-      {{on 'input' this.onTitleChange}}
-    />
-    <label for='floatingPassword'>Explain problem</label>
-  </div>
-
-  <div class='d-grid gap-2'>
-    <button
-      type='button'
-      class='btn btn-success btn-lg mt-5'
-      disabled={{this.hasEmptyField}}
-      {{on 'click' this.onCreatNewRequest}}
-    >Create request</button>
-  </div>
-</div>
-
-{{outlet}}
+<Providers::OpenRequest />
